Move Tic Tac Toe into one file
Deleted Board.py and put the class inside tictactoe.py
